+++ Sources/MarkersExtractor/Export/ExportFolderFormat.swift
@@ -1,0 +1,33 @@
+//
+//  ExportFolderFormat.swift
+//  MarkersExtractor • https://github.com/TheAcharya/MarkersExtractor
+//  Licensed under MIT License
+//
+
+import Foundation
+
+public enum ExportFolderFormat: String, CaseIterable {
+    case short
+    case medium
+    case long
+}
+
+extension ExportFolderFormat {
+    func folderName(projectName: String, profile: ExportProfileFormat) -> String {
+        switch self {
+        case .short:
+            return "\(projectName)"
+        case .medium:
+            return "\(projectName) \(nowTimestamp())"
+        case .long:
+            return "\(projectName) \(nowTimestamp()) [\(profile.name)]"
+        }
+    }
+    
+    private func nowTimestamp() -> String {
+        let now = Date()
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd hh-mm-ss"
+        return formatter.string(from: now)
+    }
+}